sam/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
bcrypt==4.3.0
beam-client==0.2.174
beta9==0.1.228
betterproto-beta9==2.0.1
black==25.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
croniter==3.0.4
cryptography==45.0.5
fastapi==0.116.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
grpcio==1.69.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
isort==5.13.2
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
paramiko==3.5.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==4.25.8
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyNaCl==1.5.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
rich==13.9.4
segment-anything==1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
typeguard==2.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
